// src/type/Solution_schedule.type.ts
export interface Solution_carrier_order {
    s_id: number;
    order_id: number;
    penalty_cost: number;
    reward: number;
}

export interface Solution_carrier_orderState {
    result: Solution_carrier_order
    loading: boolean
    error: string | null
}